server/requirements.txt
aiofiles==0.6.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
biopython==1.78
blinker==1.4
click==7.1.2
construct==2.10.63
coverage==5.5
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
editdistance==0.5.3
Flask==1.1.2
Flask-Cors==3.0.10
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kaleido==0.1.0
kiwisolver==1.3.1
line-profiler==3.2.6
llvmlite==0.36.0
logomaker==0.8
MarkupSafe==1.1.1
matplotlib==3.3.4
nbformat==5.1.2
numba==0.53.0
numpy==1.20.3
packaging==20.9
pandarallel==1.5.2
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip-autoremove==0.9.1
pip-licenses==3.3.1
plotly==4.14.3
pluggy==0.13.1
priority==1.3.0
prompt-toolkit==3.0.16
PTable==0.9.2
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
Pygments==2.8.0
pynndescent==0.5.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
Quart==0.14.1
Quart-CORS==0.4.0
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlescapy==1.0.1
tabulate==0.8.9
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
types-PyYAML==5.4.3
umap-learn==0.5.1
wcwidth==0.2.5
Werkzeug==1.0.1
wsproto==1.0.0
